Rename VerificationResult to MatchResult
